♻️ Refactor: 기존 커스텀 화면 관련 파일 삭제
Related to: #357
