--- app/src/main/java/com/mongmong/namo/presentation/ui/custom/CustomFontFragment.kt
@@ -1,15 +1,0 @@
-package com.mongmong.namo.presentation.ui.custom
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.mongmong.namo.R
-import com.mongmong.namo.databinding.FragmentCustomFontBinding
-import com.mongmong.namo.presentation.config.BaseFragment
-
-class CustomFontFragment : BaseFragment<FragmentCustomFontBinding>(R.layout.fragment_custom_font) {
-    override fun setup() {
-    }
-}
