--- app/src/main/java/com/mongmong/namo/presentation/ui/custom/CustomMyFragment.kt
@@ -1,25 +1,0 @@
-package com.mongmong.namo.presentation.ui.custom
-
-import androidx.recyclerview.widget.LinearLayoutManager
-import com.mongmong.namo.R
-import com.mongmong.namo.domain.model.Palette
-import com.mongmong.namo.databinding.FragmentCustomMyBinding
-import com.mongmong.namo.presentation.config.BaseFragment
-import com.mongmong.namo.presentation.enums.CategoryColor
-import com.mongmong.namo.presentation.ui.custom.adapter.PaletteRVAdapter
-
-class CustomMyFragment : BaseFragment<FragmentCustomMyBinding>(R.layout.fragment_custom_my) {
-    override fun setup() {
-        // 팔레트 페이지랑 동일한 데이터
-        val paletteDatas = arrayListOf(
-            Palette("기본 팔레트", CategoryColor.getAllHexColors())
-        )
-
-        //어댑터 연결
-        binding.customMyPaletteRv.apply {
-            adapter = PaletteRVAdapter(requireContext()).build(paletteDatas)
-            layoutManager = LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
-        }
-    }
-
-}
