--- app/src/main/res/layout/fragment_custom.xml
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:fitsSystemWindows="true">
-
-        <LinearLayout
-            android:id="@+id/custom_titlebar"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            android:padding="20dp"
-            android:orientation="horizontal">
-
-            <TextView
-                android:id="@+id/custom_title_tv"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="9"
-                android:text="@string/nav_setting"
-                style="@style/header"/>
-
-            <ImageView
-                android:id="@+id/custom_setting_iv"
-                android:layout_width="32dp"
-                android:layout_height="32dp"
-                android:padding="5dp"
-                android:src="@drawable/ic_settings"
-                android:layout_gravity="center_vertical"/>
-
-        </LinearLayout>
-
-        <com.google.android.material.tabs.TabLayout
-            android:id="@+id/custom_content_tb"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="25dp"
-            android:background="@color/white"
-            app:tabIndicatorColor="@color/main"
-            app:tabSelectedTextColor="@color/main"
-            app:tabRippleColor="@color/transparent"
-            app:tabTextAppearance="@style/tabTextAppearance"
-            app:layout_constraintTop_toBottomOf="@id/custom_titlebar"
-            app:tabTextColor="@color/main_text"/>
-
-        <androidx.viewpager2.widget.ViewPager2
-            android:id="@+id/custom_content_vp"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            app:layout_constraintTop_toBottomOf="@id/custom_content_tb"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"/>
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</layout>
